Move tensor(Half) from Tensor.Factories.cs to tensor_Half.cs
