+++ src/TorchSharp/Tensor/Factories/Tensor.Factories.cs
@@ -2,11 +2,10 @@
 #nullable enable
 using System;
 using System.Collections.Concurrent;
-using System.Collections.Generic;
 using System.Diagnostics.Contracts;
 using System.Linq;
 using System.Runtime.InteropServices;
-using Utils;
+using TorchSharp.Utils;
 using static TorchSharp.PInvoke.NativeMethods;
 
 namespace TorchSharp
@@ -71,42 +70,6 @@ namespace TorchSharp
         public static Tensor arange(Scalar stop, ScalarType? dtype = null, Device? device = null, bool requires_grad = false)
         {
             return arange(0, stop, 1, dtype, device, requires_grad);
-        }
-
-        /// <summary>
-        /// Create a new tensor filled with a given value
-        /// </summary>
-        private static Tensor _full(ReadOnlySpan<long> size, Scalar value, ScalarType? dtype = null, Device? device = null, bool requires_grad = false, string[]? names = null)
-        {
-            device = InitializeDevice(device);
-            if (!dtype.HasValue) {
-                // Determine the element type dynamically.
-                if (value.Type.IsIntegral()) {
-                    dtype = ScalarType.Int64;
-                } else {
-                    dtype = get_default_dtype();
-                }
-            }
-
-            unsafe {
-                fixed (long* psizes = size) {
-                    var handle = THSTensor_full((IntPtr)psizes, size.Length, value.Handle, (sbyte)dtype, (int)device.type, device.index, requires_grad);
-                    if (handle == IntPtr.Zero) {
-                        GC.Collect();
-                        GC.WaitForPendingFinalizers();
-                        handle = THSTensor_full((IntPtr)psizes, size.Length, value.Handle, (sbyte)dtype, (int)device.type, device.index, requires_grad);
-                    }
-                    if (handle == IntPtr.Zero) { CheckForErrors(); }
-                    var result = new Tensor(handle);
-
-                    if (names != null && names.Length > 0) {
-
-                        result.rename_(names);
-                    }
-
-                    return result;
-                }
-            }
         }
 
         /// <summary>
@@ -211,111 +174,7 @@ namespace TorchSharp
                 }
             }
         }
-#if NET6_0_OR_GREATER
-        /// <summary>
-        /// Create a tensor from an array of values, shaping it based on the shape passed in.
-        /// </summary>
-        /// <remarks>The Torch runtime does not take ownership of the data, so there is no device argument.</remarks>
-        [Pure]
-        public static Tensor tensor(IList<Half> rawArray, ReadOnlySpan<long> dimensions, ScalarType? dtype = null, Device? device = null, bool requires_grad = false, string[]? names = null)
-        {
-            return _tensor_generic(rawArray.ToArray(), dimensions, (sbyte)ScalarType.Float16, dtype, device, requires_grad, false, names);
-        }
 
-        /// <summary>
-        /// Create a tensor from an array of values, shaping it based on the shape passed in.
-        /// </summary>
-        [Pure]
-        public static Tensor tensor(Half[] rawArray, ScalarType? dtype = null, Device? device = null, bool requires_grad = false, string[]? names = null)
-        {
-            return _tensor_generic(rawArray, stackalloc long[] { rawArray.LongLength }, (sbyte)ScalarType.Float16, dtype, device, requires_grad, names: names);
-        }
-
-        /// <summary>
-        /// Create a tensor from an array of values, shaping it based on the shape passed in.
-        /// </summary>
-        [Pure]
-        public static Tensor tensor(Half[] rawArray, ReadOnlySpan<long> dimensions, ScalarType? dtype = null, Device? device = null, bool requires_grad = false, string[]? names = null)
-        {
-            return _tensor_generic(rawArray, dimensions, (sbyte)ScalarType.Float16, dtype, device, requires_grad, names: names);
-        }
-
-        /// <summary>
-        /// Create a 1-D tensor from an array of values, shaping it based on the input array.
-        /// </summary>
-        /// <remarks>The Torch runtime does not take ownership of the data, so there is no device argument.</remarks>
-        [Pure]
-        public static Tensor tensor(IList<Half> rawArray, ScalarType? dtype = null, Device? device = null, bool requires_grad = false, string[]? names = null)
-        {
-            return tensor(rawArray, stackalloc long[] { (long)rawArray.Count }, dtype, device, requires_grad, names: names);
-        }
-
-        /// <summary>
-        /// Create a tensor from an array of values, organizing it as a two-dimensional tensor.
-        /// </summary>
-        /// <remarks>
-        /// The Torch runtime does not take ownership of the data, so there is no device argument.
-        /// The input array must have rows * columns elements.
-        /// </remarks>
-        [Pure]
-        public static Tensor tensor(IList<Half> rawArray, long rows, long columns, ScalarType? dtype = null, Device? device = null, bool requires_grad = false, string[]? names = null)
-        {
-            return tensor(rawArray, stackalloc long[] { rows, columns }, dtype, device, requires_grad, names: names);
-        }
-
-        /// <summary>
-        /// Create a tensor from an array of values, organizing it as a three-dimensional tensor.
-        /// </summary>
-        /// <remarks>
-        /// The Torch runtime does not take ownership of the data, so there is no device argument.
-        /// The input array must have dim0*dim1*dim2 elements.
-        /// </remarks>
-        [Pure]
-        public static Tensor tensor(IList<Half> rawArray, long dim0, long dim1, long dim2, ScalarType? dtype = null, Device? device = null, bool requires_grad = false, string[]? names = null)
-        {
-            return tensor(rawArray, stackalloc long[] { dim0, dim1, dim2 }, dtype, device, requires_grad, names: names);
-        }
-
-        /// <summary>
-        /// Create a tensor from an array of values, organizing it as a four-dimensional tensor.
-        /// </summary>
-        /// <remarks>
-        /// The Torch runtime does not take ownership of the data, so there is no device argument.
-        /// The input array must have dim0*dim1*dim2*dim3 elements.
-        /// </remarks>
-        [Pure]
-        public static Tensor tensor(IList<Half> rawArray, long dim0, long dim1, long dim2, long dim3, ScalarType? dtype = null, Device? device = null, bool requires_grad = false, string[]? names = null)
-        {
-            return tensor(rawArray, stackalloc long[] { dim0, dim1, dim2, dim3 }, dtype, device, requires_grad, names: names);
-        }
-
-        /// <summary>
-        /// Create a two-dimensional tensor from a two-dimensional array of values.
-        /// </summary>
-        [Pure]
-        public static Tensor tensor(Half[,] rawArray, ScalarType? dtype = null, Device? device = null, bool requires_grad = false, string[]? names = null)
-        {
-            return _tensor_generic(rawArray, stackalloc long[] { rawArray.GetLongLength(0), rawArray.GetLongLength(1) }, (sbyte)ScalarType.Float16, dtype, device, requires_grad, names: names);
-        }
-
-        /// <summary>
-        /// Create a three-dimensional tensor from a three-dimensional array of values.
-        /// </summary>
-        [Pure]
-        public static Tensor tensor(Half[,,] rawArray, ScalarType? dtype = null, Device? device = null, bool requires_grad = false, string[]? names = null)
-        {
-            return _tensor_generic(rawArray, stackalloc long[] { rawArray.GetLongLength(0), rawArray.GetLongLength(1), rawArray.GetLongLength(2) }, (sbyte)ScalarType.Float16, dtype, device, requires_grad, names: names);
-        }
-
-        /// <summary>
-        /// Create a four-dimensional tensor from a four-dimensional array of values.
-        /// </summary>
-        [Pure]
-        public static Tensor tensor(Half[,,,] rawArray, ScalarType? dtype = null, Device? device = null, bool requires_grad = false, string[]? names = null)
-        {
-            return _tensor_generic(rawArray, stackalloc long[] { rawArray.GetLongLength(0), rawArray.GetLongLength(1), rawArray.GetLongLength(2), rawArray.GetLongLength(3) }, (sbyte)ScalarType.Float16, dtype, device, requires_grad, names: names);
-        }
-#endif
         /// <summary>
         /// Creates a <see cref="torch.Tensor">torch tensor</see> from an arbitrary <see cref="Array">array</see>.
         /// </summary>
